Replaced EvaluationFunctor with FunctionContainer
